refactor control modules into own files
