+++ lib/constants.py
@@ -21,6 +21,7 @@ class PreparedBlockData:
     compression_type: int
     should_erase: bool
     uds_checksum: bytes
+    block_name: str
 
     def __init__(
         self,
@@ -31,6 +32,7 @@ class PreparedBlockData:
         compression_type,
         should_erase,
         uds_checksum,
+        block_name,
     ):
         self.block_number = block_number
         self.block_encrypted_bytes = block_bytes
@@ -39,6 +41,7 @@ class PreparedBlockData:
         self.compression_type = compression_type
         self.should_erase = should_erase
         self.uds_checksum = uds_checksum
+        self.block_name = block_name
 
 
 class ChecksumState(Enum):
@@ -69,13 +72,6 @@ class ControlModuleIdentifier:
 
 
 ecu_control_module_identifier = ControlModuleIdentifier(0x7E8, 0x7E0)
-dsg_control_module_identifier = ControlModuleIdentifier(0x7E9, 0x7E1)
-
-# In Simos, we can send the maximum allowable size worth of compressed data in an ISO-TP request when we are using the "normal" TransferData system.
-# For DSG, things seem a little iffier so we use smaller block sizes.
-
-block_transfer_sizes_simos = {1: 0xFFD, 2: 0xFFD, 3: 0xFFD, 4: 0xFFD, 5: 0xFFD}
-block_transfer_sizes_dsg = {2: 0x4B0, 3: 0x800, 4: 0x800}
 
 
 class FlashInfo:
@@ -142,364 +138,6 @@ def internal_path(*path_parts) -> str:
             os.path.join(os.getcwd(), os.path.dirname(__file__))
         )
         return os.path.join(__location__, os.path.pardir, *path_parts)
-
-
-software_version_location_dsg = {2: [0x0, 0x0], 3: [0x0, 0x0], 4: [0x1FFE0, 0x1FFE4]}
-
-box_code_location_dsg = {2: [0x0, 0x0], 3: [0x0, 0x0], 4: [0x1FFC0, 0x1FFD3]}
-
-block_identifiers_dsg = {2: 0x30, 3: 0x50, 4: 0x51}
-
-# DSG uses external UDS checksum only for the Driver block, the other blocks are internally checksummed.
-# See dsg_checksum.py for an implementation.
-block_checksums_dsg = {
-    2: bytes.fromhex("F974176E"),
-    3: bytes.fromhex("FFFFFFFF"),
-    4: bytes.fromhex("FFFFFFFF"),
-}
-
-block_lengths_dsg = {
-    2: 0x80E,  # DRIVER
-    3: 0x130000,  # ASW
-    4: 0x20000,  # CAL
-}
-
-dsg_sa2_script = bytes.fromhex(
-    "68028149680593A55A55AA4A0587810595268249845AA5AA558703F780384C"
-)
-block_names_frf_dsg = {2: "FD_2", 3: "FD_3", 4: "FD_4"}
-
-dsg_flash_info = FlashInfo(
-    None,
-    block_lengths_dsg,
-    dsg_sa2_script,
-    None,
-    None,
-    None,
-    block_names_frf_dsg,
-    None,
-    None,
-    None,
-    block_identifiers_dsg,
-    block_checksums_dsg,
-    dsg_control_module_identifier,
-    software_version_location_dsg,
-    box_code_location_dsg,
-    block_transfer_sizes_dsg,
-)
-
-# When we're performing WriteWithoutErase, we need to write 8 bytes at a time in "patch areas" to allow the ECC operation to be performed correctly across the patched data.
-# But, when we're just "writing" 0s (which we can't actually do), we can go faster and fill an entire 256-byte Assembly Page in the flash controller as ECC will not work anyway.
-# Internally, we're basically stuffing the Assembly Page for the flash controller and the method return does not wait for controller readiness, so we will also need to resend data repeatedly.
-
-
-def s18_block_transfer_sizes_patch(block_number: int, address: int) -> int:
-    if block_number != 4:
-        print(
-            "Only patching H__0001's Block 4 / ASW3 using a provided patch is supported at this time! If you have a patch for another block, please fill in its data areas here."
-        )
-        exit()
-    if address < 0x9600:
-        return 0x100
-    if address >= 0x9600 and address < 0x9800:
-        return 0x8
-    if address >= 0x9800 and address < 0x7DD00:
-        return 0x100
-    if address >= 0x7DD00 and address < 0x7E200:
-        return 0x8
-    if address >= 0x7E200 and address < 0x7F900:
-        return 0x100
-    return 0x8
-
-
-def s1810_block_transfer_sizes_patch(block_number: int, address: int) -> int:
-    if block_number != 2:
-        print(
-            "Only patching Q__0005's Block 2 / ASW1 using a provided patch is supported at this time! If you have a patch for another block, please fill in its data areas here."
-        )
-        exit()
-    if address < 0x5CB00:
-        return 0x100
-    if address >= 0x5CB00 and address < 0x5CC00:
-        return 0x8
-    if address >= 0x5CC00 and address < 0xB3000:
-        return 0x100
-    if address >= 0xB3000 and address < 0xB3100:
-        return 0x8
-    if address >= 0xB3100 and address < 0xDFB00:
-        return 0x100
-    return 0x8
-
-
-software_version_location_simos = {
-    1: [0x437, 0x43F],
-    2: [0x627, 0x62F],
-    3: [0x203, 0x20B],
-    4: [0x203, 0x20B],
-    5: [0x23, 0x2B],
-    7: [0, 0],
-    9: [0, 0],
-}
-
-box_code_location_simos = {
-    1: [0x0, 0x0],
-    2: [0x0, 0x0],
-    3: [0x0, 0x0],
-    4: [0x0, 0x0],
-    5: [0x60, 0x6B],
-    7: [0, 0],
-    9: [0x0, 0x0],
-}
-
-block_names_frf_s18 = {1: "FD_0", 2: "FD_1", 3: "FD_2", 4: "FD_3", 5: "FD_4"}
-
-block_identifiers_simos = {1: 1, 2: 2, 3: 3, 4: 4, 5: 5}
-
-# Simos does not use block checksums sent over UDS but rather checksums internally. See checksum.py for the internal checksum implementation.
-block_checksums_simos = {
-    1: bytes.fromhex("00000000"),
-    2: bytes.fromhex("00000000"),
-    3: bytes.fromhex("00000000"),
-    4: bytes.fromhex("00000000"),
-    5: bytes.fromhex("00000000"),
-    7: bytes.fromhex("00000000"),
-    9: bytes.fromhex("00000000"),
-}
-
-# Simos12 Flash Info
-
-# block sizes for S12
-block_lengths_s12 = {
-    1: 0x1FE00,  # CBOOT
-    2: 0xBFC00,  # ASW1
-    3: 0xBFC00,  # ASW2
-    4: 0xBFC00,  # ASW3
-    5: 0x6FC00,  # CAL
-    6: 0x1FE00,  # CBOOT_temp
-}
-
-# The base address of each block on simos12
-
-base_addresses_s12 = {
-    0: 0x80000000,  # SBOOT
-    1: 0x80020000,  # CBOOT
-    2: 0x800C0000,  # ASW1
-    3: 0x80180000,  # ASW2
-    4: 0x80240000,  # ASW3
-    5: 0xA0040000,  # CAL
-    6: 0x80080000,  # CBOOT_temp
-}
-
-s12_iv = bytes.fromhex("306e37426b6b536f316d4a6974366d34")
-s12_key = bytes.fromhex("314d7536416e3047396a413252356f45")
-
-s12_sa2_script = bytes.fromhex(
-    "6803814A10680393290720094A05872212195482499309011953824A058730032009824A0181494C"
-)
-
-s12_flash_info = FlashInfo(
-    base_addresses_s12,
-    block_lengths_s12,
-    s12_sa2_script,
-    s12_key,
-    s12_iv,
-    s18_block_transfer_sizes_patch,
-    block_names_frf_s18,
-    "",
-    0,
-    "",
-    block_identifiers_simos,
-    block_checksums_simos,
-    ecu_control_module_identifier,
-    software_version_location_simos,
-    box_code_location_simos,
-    block_transfer_sizes_simos,
-)
-
-# Simos18.1 / 18.6 Flash Info
-
-# The base address of each block for S18.1
-base_addresses_s18 = {
-    0: 0x80000000,  # SBOOT
-    1: 0x8001C000,  # CBOOT
-    2: 0x80040000,  # ASW1
-    3: 0x80140000,  # ASW2
-    4: 0x80880000,  # ASW3
-    5: 0xA0800000,  # CAL
-    6: 0x80840000,  # CBOOT_temp
-}
-
-# The size of each block
-block_lengths_s18 = {
-    1: 0x23E00,  # CBOOT
-    2: 0xFFC00,  # ASW1
-    3: 0xBFC00,  # ASW2
-    4: 0x7FC00,  # ASW3
-    5: 0x7FC00,  # CAL
-    6: 0x23E00,  # CBOOT_temp
-}
-
-s18_key = bytes.fromhex("98D31202E48E3854F2CA561545BA6F2F")
-s18_iv = bytes.fromhex("E7861278C508532798BCA4FE451D20D1")
-
-sa2_script_s18 = bytes.fromhex(
-    "6802814A10680493080820094A05872212195482499307122011824A058703112010824A0181494C"
-)
-
-s18_flash_info = FlashInfo(
-    base_addresses_s18,
-    block_lengths_s18,
-    sa2_script_s18,
-    s18_key,
-    s18_iv,
-    s18_block_transfer_sizes_patch,
-    block_names_frf_s18,
-    "8V0906259H",
-    4,
-    internal_path("docs", "patch.bin"),
-    block_identifiers_simos,
-    block_checksums_simos,
-    ecu_control_module_identifier,
-    software_version_location_simos,
-    box_code_location_simos,
-    block_transfer_sizes_simos,
-)
-
-# Simos 18.10 Flash Info
-
-block_names_frf_s1810 = {
-    1: "FD_01DATA",
-    2: "FD_02DATA",
-    3: "FD_03DATA",
-    4: "FD_04DATA",
-    5: "FD_05DATA",
-}
-
-base_addresses_s1810 = {
-    0: 0x80000000,  # SBOOT
-    1: 0x80800000,  # CBOOT
-    2: 0x80020000,  # ASW1
-    3: 0x80100000,  # ASW2
-    4: 0x808C0000,  # ASW3
-    5: 0xA0820000,  # CAL
-    6: 0x80880000,  # CBOOT_temp
-}
-
-# The size of each block
-block_lengths_s1810 = {
-    1: 0x1FE00,  # CBOOT
-    2: 0xDFC00,  # ASW1
-    3: 0xFFC00,  # ASW2
-    4: 0x13FC00,  # ASW3
-    5: 0x9FC00,  # CAL
-    6: 0x1FE00,  # CBOOT_temp
-}
-
-s1810_key = bytes.fromhex("AE540502E48E3854DBCA1A1545BA6F33")
-s1810_iv = bytes.fromhex("62F313FA5C08532798BCA452471D20D5")
-
-sa2_script_s1810 = bytes.fromhex(
-    "6803814A10680293050520154A058722121954824993F423BF7D824A05875A63FC5E824A0181494C"
-)
-
-s1810_flash_info = FlashInfo(
-    base_addresses_s1810,
-    block_lengths_s1810,
-    sa2_script_s1810,
-    s1810_key,
-    s1810_iv,
-    s1810_block_transfer_sizes_patch,
-    block_names_frf_s1810,
-    "5G0906259Q",
-    2,
-    internal_path("docs", "patch_1810.bin"),
-    block_identifiers_simos,
-    block_checksums_simos,
-    ecu_control_module_identifier,
-    software_version_location_simos,
-    box_code_location_simos,
-    block_transfer_sizes_simos,
-)
-
-# Simos 18.41 Flash Info
-
-block_names_frf_s1841 = {
-    1: "FD_01FLASHDATA",
-    2: "FD_02FLASHDATA",
-    3: "FD_03FLASHDATA",
-    4: "FD_04FLASHDATA",
-    5: "FD_05FLASHDATA",
-}
-
-s1841_key = bytes.fromhex("6E3FE03619F138798CB4ECDCC762005F")
-s1841_iv = bytes.fromhex("000102030405060708090A0B0C0D0E0F")
-
-sa2_script_s1841 = bytes.fromhex(
-    "6802814A10680493C1387FA34A05872212195482499318102012824A058728051977824A0181494C"
-)
-
-s1841_flash_info = FlashInfo(
-    base_addresses_s1810,
-    block_lengths_s1810,
-    sa2_script_s1841,
-    s1841_key,
-    s1841_iv,
-    None,
-    block_names_frf_s1841,
-    None,
-    None,
-    None,
-    block_identifiers_simos,
-    block_checksums_simos,
-    ecu_control_module_identifier,
-    software_version_location_simos,
-    box_code_location_simos,
-    block_transfer_sizes_simos,
-)
-
-
-# The location of each checksum in the bin
-checksum_block_location = {
-    0: 0x300,  # SBOOT
-    1: 0x300,  # CBOOT
-    2: 0x300,  # ASW1
-    3: 0x0,  # ASW2
-    4: 0x0,  # ASW3
-    5: 0x300,  # CAL
-    6: 0x340,  # CBOOT_temp
-}
-
-# The location of the addresses for ECM3 Level 2 CAL monitoring
-# 'Early' cars seem to have a different version of the ECM3 module which looks in a different place for the ECM2 Calibration offsets to checksum
-# 'Early' cars also calculate the offsets in a different way.
-
-ecm3_cal_monitor_addresses_early = 0x540  # Offset into ASW1
-ecm3_cal_monitor_addresses = 0x520  # Offset into ASW1
-ecm3_cal_monitor_offset_uncached = 0
-ecm3_cal_monitor_offset_cached = 0x20000000
-ecm3_cal_monitor_checksum = 0x400  # Offset into CAL
-
-# Conversion dict for block name to number
-block_name_to_int = {
-    "CBOOT": 1,
-    "ASW1": 2,
-    "ASW2": 3,
-    "ASW3": 4,
-    "CAL": 5,
-    "CBOOT_TEMP": 6,
-    "PATCH_ASW1": 7,
-    "PATCH_ASW2": 8,
-    "PATCH_ASW3": 9,
-}
-
-int_to_block_name = dict((reversed(item) for item in block_name_to_int.items()))
-
-
-def block_to_number(blockname: str) -> int:
-    if blockname.isdigit():
-        return int(blockname)
-    else:
-        return block_name_to_int[blockname.upper()]
 
 
 data_records: List[DataRecord] = [
