+++ lib/modules/__init.py__
@@ -1,0 +1,6 @@
+import simosshared
+import simos12
+import simos18
+import simos1810
+import simos184
+import dq250mqb
